Refactoring: planning-scene-server is now separate into ros-agnostic class and ros-specific wrapper.
