+++ knowrob_moveit_core/include/knowrob_moveit_core/knowrob_moveit_core.hpp
@@ -31,6 +31,7 @@
 #ifndef KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP 
 #define KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP 
 
+#include <knowrob_moveit_core/planning_scene.hpp>
 #include <knowrob_moveit_core/planning_scene_server.hpp>
 #include <knowrob_moveit_core/utils.hpp>
