+++ knowrob_moveit_core/include/knowrob_moveit_core/planning_scene_server.hpp
@@ -30,17 +30,12 @@
 #ifndef KNOWROB_MOVEIT_CORE_PLANNING_SCENE_SERVER_HPP 
 #define KNOWROB_MOVEIT_CORE_PLANNING_SCENE_SERVER_HPP 
 
+#include <knowrob_moveit_core/planning_scene.hpp>
 #include <knowrob_moveit_core/utils.hpp>
-#include <moveit/planning_scene/planning_scene.h>
-#include <urdf/model.h>
-#include <sensor_msgs/JointState.h>
 #include <knowrob_moveit_msgs/CheckCollisions.h>
-#include <vector>
-#include <functional>
 
 namespace knowrob_moveit_core
 {
-  // TODO: separate into ROS-dependent and ROS-agnostic functionality
   class PlanningSceneServer
   {
     public:
@@ -58,60 +53,7 @@ namespace knowrob_moveit_core
     private:
       ros::NodeHandle nh_;
       ros::ServiceServer service_;
-      size_t urdf_hash_, srdf_hash_;
-  
-      planning_scene::PlanningScenePtr ps_ptr_;
-  
-      // TODO: see whether it is also fine to just remember the robot model
-      void createPlanningScene(const std::string& urdf, const std::string& srdf)
-      {
-        ROS_INFO("Hashing started.");
-        std::hash<std::string> hash_fn;
-        size_t new_urdf_hash = hash_fn(urdf);
-        size_t new_srdf_hash = hash_fn(srdf);
-        ROS_INFO("Hashing done.");
-  
-        if(!ps_ptr_ || urdf_hash_ != new_urdf_hash || srdf_hash_ != new_srdf_hash)
-        {
-          ROS_INFO("Parsing urdf started.");
-          boost::shared_ptr<urdf::Model> urdf_ptr(new urdf::Model());
-          if (!urdf_ptr->initString(urdf))
-            throw std::runtime_error("Could not parse given urdf. Aborting.");
-          ROS_INFO("Parsing urdf done.");
-  
-          ROS_INFO("Parsing srdf started.");
-          boost::shared_ptr<srdf::Model> srdf_ptr(new srdf::Model());
-          if (!srdf_ptr->initString(*urdf_ptr, srdf))
-            throw std::runtime_error("Could not parse given srdf. Aborting.");
-          ROS_INFO("Parsing srdf done.");
-     
-          ROS_INFO("Creation planning scene started.");
-          boost::shared_ptr<const srdf::Model> srdf_const_ptr = srdf_ptr;
-          boost::shared_ptr<const urdf::Model> urdf_const_ptr = urdf_ptr;
-      
-          ps_ptr_ = planning_scene::PlanningScenePtr(new planning_scene::PlanningScene(urdf_const_ptr, srdf_const_ptr));
-          urdf_hash_ = new_urdf_hash;
-          srdf_hash_ = new_srdf_hash;
-          ROS_INFO("Creation planning scene done.");
-        }
-      }
-  
-      std::vector<moveit_msgs::ContactInformation> calculate_collisions()
-      {
-        ROS_INFO("KnowRob-MoveIt check_collisions started.");
-      
-        collision_detection::CollisionRequest collision_request;
-        collision_detection::CollisionResult collision_result;
-        // TODO: check whether those are useful defaults
-        collision_request.contacts = true;
-        collision_request.max_contacts = 10;
-        collision_result.clear();
-        ps_ptr_->checkCollision(collision_request, collision_result);
-      
-        ROS_INFO("KnowRob-MoveIt check_collisions finished.");
-      
-        return collisionResultToMsg(collision_result);
-      }
+      PlanningScene planning_scene_;
 
       bool callback(knowrob_moveit_msgs::CheckCollisions::Request& request, 
           knowrob_moveit_msgs::CheckCollisions::Response& response)
@@ -119,23 +61,17 @@ namespace knowrob_moveit_core
         ROS_INFO("KnowRob-MoveIt check_collisions called.");
       
         try{
-          createPlanningScene(request.urdf_model, request.srdf_model);
+          response.contacts = planning_scene_.checkCollisions(request.urdf_model, request.srdf_model,
+              request.joint_states);
         }
         catch(const std::exception& e)
         {
           ROS_ERROR("%s", e.what());
           return false;
         }
-  
-        ROS_INFO("Update planning scene started.");
-        moveit_msgs::PlanningScene diff_ps_msg;
-        diff_ps_msg.is_diff = true;
-        diff_ps_msg.robot_state.joint_state = request.joint_states;
-        ps_ptr_->setPlanningSceneDiffMsg(diff_ps_msg);
-        ROS_INFO("Update planning scene done.");
-     
-        response.contacts = calculate_collisions();
-      
+
+        ROS_INFO("KnowRob-MoveIt check_collisions finished.");
+ 
         return true;
       }
